# docs/swagger.yaml
basePath: /
definitions:
  core.TipsetState:
    properties:
      description:
        type: string
      not_found_state:
        type: integer
      state:
        type: integer
      tipset:
        type: integer
      topic:
        type: string
      version:
        type: integer
    required:
    - topic
    type: object
  core.Topic:
    properties:
      topic:
        example: messages/vm_messages...
        type: string
    required:
    - topic
    type: object
  utils.ResponseWithRequestId:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
    type: object
host: extractor-api.spacescope.io
info:
  contact:
    email: xueyou@starboardventures.io
    name: xueyouchen
  description: spacescope data extraction api backend
  termsOfService: http://swagger.io/terms/
  title: spacescope data extraction notify backend
  version: "1.0"
paths:
  /api/v1/gapfill:
    post:
      consumes:
      - application/json
      - application/json
      description: automatic fill the gap's tipsets.
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
      tags:
      - DATA-EXTRACTION-API-Internal-V1-CallByManual
  /api/v1/ping:
    get:
      consumes:
      - application/json
      - application/json
      description: Healthy examination
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: pong
          schema:
            type: string
        "500":
          description: error:...
          schema:
            type: string
      tags:
      - Sys
  /api/v1/task_state:
    post:
      consumes:
      - application/json
      - application/json
      description: task will report tipset state with this API.
      parameters:
      - description: TipsetState
        in: body
        name: TipsetState
        schema:
          $ref: '#/definitions/core.TipsetState'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
      tags:
      - DATA-EXTRACTION-API-Internal-V1-CallByTaskModel
  /api/v1/topic:
    delete:
      consumes:
      - application/json
      - application/json
      description: delete a topic.
      parameters:
      - description: Topic
        in: body
        name: Topic
        schema:
          $ref: '#/definitions/core.Topic'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
      tags:
      - DATA-EXTRACTION-API-Internal-V1-CallByTaskModel
    post:
      consumes:
      - application/json
      - application/json
      description: task group will sign in a mq topic use this API.
      parameters:
      - description: Topic
        in: body
        name: Topic
        schema:
          $ref: '#/definitions/core.Topic'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
      tags:
      - DATA-EXTRACTION-API-Internal-V1-CallByTaskModel
  /api/v1/walk:
    post:
      consumes:
      - application/json
      - application/json
      description: walk the historical DAG's tipsets.
      parameters:
      - in: query
        name: from
        type: integer
      - in: query
        name: to
        type: integer
      - in: query
        name: topic
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseWithRequestId'
      tags:
      - DATA-EXTRACTION-API-Internal-V1-CallByManual
swagger: "2.0"